Moving Serial UniqueToken to it's own file
